+++ src/EditorFeatures/Core/CommandHandlers/AbstractGoToCommandHandler.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Collections.Immutable;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Editor.Host;
@@ -18,7 +17,6 @@ using Microsoft.CodeAnalysis.Internal.Log;
 using Microsoft.CodeAnalysis.Notification;
 using Microsoft.CodeAnalysis.Shared.Extensions;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
-using Microsoft.CodeAnalysis.Shared.Utilities;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.VisualStudio.Commanding;
 using Microsoft.VisualStudio.Text;
@@ -29,7 +27,7 @@ using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.Editor.CommandHandlers
 {
-    internal abstract class AbstractGoToCommandHandler<TLanguageService, TCommandArgs> : ICommandHandler<TCommandArgs>
+    internal abstract partial class AbstractGoToCommandHandler<TLanguageService, TCommandArgs> : ICommandHandler<TCommandArgs>
         where TLanguageService : class, ILanguageService
         where TCommandArgs : EditorCommandArgs
     {
@@ -261,197 +259,6 @@ namespace Microsoft.CodeAnalysis.Editor.CommandHandlers
                 Contract.ThrowIfNull(document);
                 await FindActionAsync(service, document, position, findContext, cancellationToken).ConfigureAwait(false);
             }
-        }
-
-        /// <summary>
-        /// An impl of <see cref="IFindUsagesContext"/> that will buffer results internally to either be shown to the 
-        /// user immediately if the find command completes quickly, or which will be pushed into the streaming presenter 
-        /// if the search is taking too long.
-        /// </summary>
-        private sealed class SwappableFindUsagesContext : IFindUsagesContext, IStreamingProgressTracker
-        {
-            /// <summary>
-            /// Lock which controls access to all members below.
-            /// </summary>
-            private readonly SemaphoreSlim _gate = new(initialCount: 1);
-
-            /// <summary>
-            /// The underlying presenter context to forward messages to once the presenter is opened.  Prior to having 
-            /// this, we will buffer the results within ourselves.
-            /// </summary>
-            private IFindUsagesContext? _streamingPresenterContext;
-
-            // Values we buffer inside ourselves until _streamingPresenterContext is non-null.  Once non-null, we'll push
-            // the values into it and forward all future calls from that point to it.
-
-            private int _totalItemCount;
-            private int _itemsCompleted;
-
-            private string? _message;
-            private string? _searchTitle;
-
-            private ImmutableArray<DefinitionItem>.Builder? _definitions = ImmutableArray.CreateBuilder<DefinitionItem>();
-
-            public SwappableFindUsagesContext()
-            {
-            }
-
-            /// <summary>
-            /// Message to show the user in case of failure.
-            /// </summary>
-            public async Task<string?> GetMessageAsync(CancellationToken cancellationToken)
-            {
-                using (await _gate.DisposableWaitAsync(cancellationToken).ConfigureAwait(false))
-                {
-                    Contract.ThrowIfTrue(_streamingPresenterContext != null, "Should not be called if we've switched over to the streaming presenter");
-                    return _message;
-                }
-            }
-
-            /// <summary>
-            /// Message to show the user in case of failure.
-            /// </summary>
-            public async Task<string?> GetSearchTitleAsync(CancellationToken cancellationToken)
-            {
-                using (await _gate.DisposableWaitAsync(cancellationToken).ConfigureAwait(false))
-                {
-                    Contract.ThrowIfTrue(_streamingPresenterContext != null, "Should not be called if we've switched over to the streaming presenter");
-                    return _searchTitle;
-                }
-            }
-
-            public async Task<ImmutableArray<DefinitionItem>> GetDefinitionsAsync(CancellationToken cancellationToken)
-            {
-                using (await _gate.DisposableWaitAsync(cancellationToken).ConfigureAwait(false))
-                {
-                    Contract.ThrowIfNull(_definitions, "This should not be called if we switched over to the presenter to show results");
-                    return _definitions.ToImmutable();
-                }
-            }
-
-            public async Task SwapAsync(IFindUsagesContext presenterContext, CancellationToken cancellationToken)
-            {
-                using (await _gate.DisposableWaitAsync(cancellationToken).ConfigureAwait(false))
-                {
-                    Contract.ThrowIfTrue(_streamingPresenterContext != null, "Trying to set the presenter multiple times.");
-
-                    // Push all values we've buffered into the new presenter context.
-
-                    await presenterContext.ProgressTracker.AddItemsAsync(_totalItemCount, cancellationToken).ConfigureAwait(false);
-                    await presenterContext.ProgressTracker.ItemsCompletedAsync(_itemsCompleted, cancellationToken).ConfigureAwait(false);
-
-                    if (_searchTitle != null)
-                        await presenterContext.SetSearchTitleAsync(_searchTitle, cancellationToken).ConfigureAwait(false);
-
-                    if (_message != null)
-                        await presenterContext.ReportMessageAsync(_message, cancellationToken).ConfigureAwait(false);
-
-                    Contract.ThrowIfNull(_definitions);
-                    foreach (var definition in _definitions.ToImmutableArray())
-                        await presenterContext.OnDefinitionFoundAsync(definition, cancellationToken).ConfigureAwait(false);
-
-                    // Now swap over to the presenter being the sink for all future callbacks, and clear any buffered data.
-                    _streamingPresenterContext = presenterContext;
-
-                    _totalItemCount = -1;
-                    _itemsCompleted = -1;
-                    _searchTitle = null;
-                    _message = null;
-                    _definitions = null;
-                }
-            }
-
-            #region IStreamingProgressTracker
-
-            IStreamingProgressTracker IFindUsagesContext.ProgressTracker => this;
-
-            async ValueTask IStreamingProgressTracker.AddItemsAsync(int count, CancellationToken cancellationToken)
-            {
-                using (await _gate.DisposableWaitAsync(cancellationToken).ConfigureAwait(false))
-                {
-                    if (_streamingPresenterContext != null)
-                    {
-                        await _streamingPresenterContext.ProgressTracker.AddItemsAsync(count, cancellationToken).ConfigureAwait(false);
-                    }
-                    else
-                    {
-                        _totalItemCount += count;
-                    }
-                }
-            }
-
-            async ValueTask IStreamingProgressTracker.ItemsCompletedAsync(int count, CancellationToken cancellationToken)
-            {
-                using (await _gate.DisposableWaitAsync(cancellationToken).ConfigureAwait(false))
-                {
-                    if (_streamingPresenterContext != null)
-                    {
-                        await _streamingPresenterContext.ProgressTracker.ItemsCompletedAsync(count, cancellationToken).ConfigureAwait(false);
-                    }
-                    else
-                    {
-                        _itemsCompleted += count;
-                    }
-                }
-            }
-
-            #endregion
-
-            #region IFindUsagesContext
-
-            async ValueTask IFindUsagesContext.ReportMessageAsync(string message, CancellationToken cancellationToken)
-            {
-                using (await _gate.DisposableWaitAsync(cancellationToken).ConfigureAwait(false))
-                {
-                    if (_streamingPresenterContext != null)
-                    {
-                        await _streamingPresenterContext.ReportMessageAsync(message, cancellationToken).ConfigureAwait(false);
-                    }
-                    else
-                    {
-                        _message = message;
-                    }
-                }
-            }
-
-            async ValueTask IFindUsagesContext.SetSearchTitleAsync(string title, CancellationToken cancellationToken)
-            {
-                using (await _gate.DisposableWaitAsync(cancellationToken).ConfigureAwait(false))
-                {
-                    if (_streamingPresenterContext != null)
-                    {
-                        await _streamingPresenterContext.SetSearchTitleAsync(title, cancellationToken).ConfigureAwait(false);
-                    }
-                    else
-                    {
-                        _searchTitle = title;
-                    }
-                }
-            }
-
-            async ValueTask IFindUsagesContext.OnDefinitionFoundAsync(DefinitionItem definition, CancellationToken cancellationToken)
-            {
-                using (await _gate.DisposableWaitAsync(cancellationToken).ConfigureAwait(false))
-                {
-                    if (_streamingPresenterContext != null)
-                    {
-                        await _streamingPresenterContext.OnDefinitionFoundAsync(definition, cancellationToken).ConfigureAwait(false);
-                    }
-                    else
-                    {
-                        Contract.ThrowIfNull(_definitions);
-                        _definitions.Add(definition);
-                    }
-                }
-            }
-
-            ValueTask IFindUsagesContext.OnReferenceFoundAsync(SourceReferenceItem reference, CancellationToken cancellationToken)
-            {
-                // Entirely ignored.  These features do not show references.
-                return ValueTaskFactory.CompletedTask;
-            }
-
-            #endregion
         }
     }
 }
